Back_v2/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.29
pydantic==2.6.4
passlib[bcrypt]==1.7.4
python-jose==3.3.0
reportlab==4.1.0
